WIP refactor modul-component webapp
