+++ tests/app/app-frame/app-frame.html
@@ -1,0 +1,10 @@
+<div>
+    <header class="mtest-header m-u--box-shadow--s">
+        <h1 class="m-u--h4 m-u--no-margin-top">MODUL <i class="m-u--font-size--s m-u--no-margin-top"><m-link url="/" >Components</m-link> - <m-link
+            mode="Link" target="_blank" url="https://ulaval.github.io/modul/">Documentation</m-link>
+            </i> </h1>
+    </header>
+    <main class="mtest-content">
+            <router-view></router-view>
+    </main>
+</div>
