+++ tests/app/index.html
@@ -4,41 +4,12 @@
 <head>
     <title>MODUL - Components</title>
     <meta name="viewport" content="width=device-width, height=device-height, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.3, user-scalable=yes" />
-    <style>
-        .mtest-header {
-            position: fixed;
-            top: 0;
-            right: 0;
-            left: 0;
-            z-index: 2;
-            padding: 16px 32px;
-            background: #fff;
-        }
-
-        .m-flex-template__menu,
-        .m-flex-template__page {
-            padding-top: 52px;
-        }
-
-        .mtest-header h1 {
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-        }
-
-        @media (max-width: 767px) {
-            .mtest-header {
-                padding: 16px;
-            }
-        }
-    </style>
 </head>
 
 <body class="m-u--app-body">
-    <header class="mtest-header m-u--box-shadow--s">
-        <h1 class="m-u--h4 m-u--no-margin-top">MODUL <i class="m-u--font-size--s m-u--no-margin-top">Components</i></h1>
-    </header>
-    <div id="vue">... Waiting</div>
+    <div class="mtest-content">
+        <div id="vue">... Waiting</div>
+    </div>
 </body>
 
 </html>
