+++ tests/app/navigation/navigation.html
@@ -1,20 +1,60 @@
 <m-flex-template padding-top="52px">
     <m-template>
-        <h1 slot="header">Menu</h1>
-        <h2>
-            <m-link url="attributes">Attributes list</m-link>
-        </h2>
-        <h2>
-            <m-link url="media-queries">Media Queries</m-link>
-        </h2>
+        <div slot="header">
+            <h1>Menu</h1>
+            <nav class="m-u--margin-top">
+                <ul class="m-navigation__submenu">
+                    <li>
+                        <m-link mode="Link" href="#component">Components</m-link>
+                    </li>
+                    <li>
+                        <m-link mode="Link" href="#directives">Directives</m-link>
+                    </li>
+                    <li>
+                        <m-link mode="Link" href="#filters">Filters</m-link>
+                    </li>
+                    <li>
+                        <m-link mode="Link" href="#mixins">Mixins</m-link>
+                    </li>
+                </ul>
+            </nav>
+        </div>
 
-        <h2>Components list</h2>
+        <a class="anchor" id="component"></a>
+        <h2>Components <m-icon name="m-svg__arrow-thin--up" size="16px" @click="scrollToTop()"></m-icon></h2>
         <nav class="m-u--margin-top">
             <ul class="m-u--bullet-list">
                 <li v-for="route in routes">
                     <m-link :url="route">{{route}}</m-link>
                 </li>
             </ul>
+        </nav>
+        <a class="anchor" id="directives"></a>
+        <h2>Directives <m-icon name="m-svg__arrow-thin--up" size="16px" @click="scrollToTop()"></m-icon>
+        </h2>
+        <nav class="m-u--margin-top">
+            <ul class="m-u--bullet-list">
+                <li v-for="route in routes">
+                    <m-link :url="route">{{route}}</m-link>
+                </li>
+            </ul>
+        </nav>
+        <a class="anchor" id="filters"></a>
+        <h2>Filters <m-icon name="m-svg__arrow-thin--up" size="16px" @click="scrollToTop()"></m-icon>
+        </h2>
+        <a class="anchor" id="mixins"></a>
+        <h2>Mixins <m-icon name="m-svg__arrow-thin--up" size="16px" @click="scrollToTop()"></m-icon>
+        </h2>
+        <nav class="m-u--margin-top">
+            <ul class="m-u--bullet-list">
+                <li>
+                    <m-link url="media-queries">Media Queries</m-link>
+                </li>
+                <li>
+                    <m-link url="attributes">Attributes list</m-link>
+                </li>
+            </ul>
+
         </nav>
 
     </m-template>
