+++ tests/app/navigation/navigation.ts
@@ -1,12 +1,17 @@
-import { ModulVue } from '../../../src/utils/vue/vue';
 import Component from 'vue-class-component';
-import WithRender from './navigation.html';
+
 import Meta from '../../../src/meta/meta';
+import { ModulVue } from '../../../src/utils/vue/vue';
+import WithRender from './navigation.html?style=./navigation.scss';
 
 @WithRender
 @Component
 export class Navigation extends ModulVue {
     public routes: string[] = [];
+    scrollToTop(): void {
+        window.scrollTo(0, 0);
+    }
+
     protected mounted(): void {
         let meta: string[] = [];
         Meta.getTags().forEach(tag => {
@@ -14,4 +19,5 @@ export class Navigation extends ModulVue {
         });
         this.routes = meta;
     }
+
 }
